New navigation database for fixes and navaids

+++ bluesky/navdb/loaddata.py
@@ -27,10 +27,12 @@ def load_coastlines():
     if not os.path.isfile(cachedir + '/coastlines.p'):
         coastvertices, coastindices = load_coastline_txt()
         with open(cachedir + '/coastlines.p', 'wb') as f:
+            print "Writing cache: coastlines.p"
             pickle.dump(coastvertices, f, pickle.HIGHEST_PROTOCOL)
             pickle.dump(coastindices, f, pickle.HIGHEST_PROTOCOL)
     else:
         with open(cachedir + '/coastlines.p', 'rb') as f:
+            print "Reading cache: coastlines.p"
             coastvertices = pickle.load(f)
             coastindices  = pickle.load(f)
 
@@ -42,6 +44,7 @@ def load_aptsurface():
         vbuf_asphalt, vbuf_concrete, vbuf_runways, vbuf_rwythr, apt_ctr_lat, apt_ctr_lon, \
             apt_indices, rwythresholds = load_aptsurface_txt()
         with open(cachedir + '/aptsurface.p', 'wb') as f:
+            print "Writing cache: aptsurface.p"
             pickle.dump(vbuf_asphalt, f, pickle.HIGHEST_PROTOCOL)
             pickle.dump(vbuf_concrete, f, pickle.HIGHEST_PROTOCOL)
             pickle.dump(vbuf_runways , f, pickle.HIGHEST_PROTOCOL)
@@ -51,8 +54,10 @@ def load_aptsurface():
             pickle.dump(apt_indices  , f, pickle.HIGHEST_PROTOCOL)
         with open(cachedir + '/rwythresholds.p', 'wb') as f:
             pickle.dump(rwythresholds,  f, pickle.HIGHEST_PROTOCOL)
+            print "Writing cache: rwythresholds.p"
     else:
         with open(cachedir + '/aptsurface.p', 'rb') as f:
+            print "Reading cache: aptsurface.p"
             vbuf_asphalt  = pickle.load(f)
             vbuf_concrete = pickle.load(f)
             vbuf_runways  = pickle.load(f)
@@ -66,17 +71,19 @@ def load_aptsurface():
 
 def load_navdata():
 
-    cache_ok = False # We still need to check cache content and version
+    cache_ok       = False # We still need to check cache content and version
+    cache_complete = False # and completeness
     
     # Check or regenerate cache
-    while not cache_ok:
+    while not (cache_ok and cache_complete):
 
         # Does cache exist or not? If not, make a new cachefile
         if not os.path.isfile(cachedir + '/navdata.p'):
 
             wptdata, aptdata, firdata, codata = load_navdata_txt()
-
+            
             with open(cachedir + '/navdata.p', 'wb') as f:
+                print "Writing cache: navdata.p"
                 pickle.dump(wptdata,      f, pickle.HIGHEST_PROTOCOL)
                 pickle.dump(aptdata,      f, pickle.HIGHEST_PROTOCOL)
                 pickle.dump(firdata,      f, pickle.HIGHEST_PROTOCOL)
@@ -91,7 +98,9 @@ def load_navdata():
 
             datecache = os.path.getmtime(cachedir + '/navdata.p')
             
-            datesource = max(os.path.getmtime(data_path + "/global/waypoints.dat"),
+            datesource = max(os.path.getmtime(data_path + "/global/nav.dat"),
+                             os.path.getmtime(data_path + "/global/fix.dat"),
+                             os.path.getmtime(data_path + "/global/awy.dat"),
                              os.path.getmtime(data_path + "/global/airports.dat"),
                              os.path.getmtime(data_path + "/global/icao-countries.dat"))
 
@@ -99,51 +108,35 @@ def load_navdata():
                 datesource = max(datesource,
                                  os.path.getmtime(data_path+"/global/fir/"+f))
 
+            # Check for cache ok
+            cache_ok = (datecache > datesource)  
 
-            # Try reading data from cache
-            with open(cachedir + '/navdata.p', 'rb') as f:
-                if datecache > datesource:
-                    wptdata = pickle.load(f)
-                    aptdata = pickle.load(f)
-                    firdata = pickle.load(f)
-                    try:
+            if cache_ok:
+                # Try reading data from cache
+                with open(cachedir + '/navdata.p', 'rb') as f:
+                    print "Reading cache: navdata.p"
+                    try:                    
+                        wptdata = pickle.load(f)
+                        aptdata = pickle.load(f)
+                        firdata = pickle.load(f)
                         codata = pickle.load(f)
                     except:
-                        codata = {}
+                        cache_ok = False
 
-                # Do not read data from an invalid file, set empty dicts
-                else:
+            
+            # Do not read data from an invalid file, set empty dicts
+            else:
                     wptdata = {}
                     aptdata = {}
                     firdata = {}
                     codata  = {}
 
-            # Check whether cached data is okay, insert any data check here
-
-            # Is requires set a subset (<=) of found set? 
-            # (so are all listed keys in there?)
+            # Integrity & completenes check: 
+            # used in case of reading more vars from same files
+            cache_complete = "wpfreq" in wptdata.keys()
 
             # If it is the same set of data:
-            if not type(wptdata)==int: # Catch int if version nr was still in data file
-                wpdata_ok  = set(["wpid","wplat","wplon","wpapt","wptype","wpco"]) \
-                                 <= set(wptdata.keys())
-            else:
-                wpdata_ok = False
-                                                                             
-            apdata_ok  = set(["apid","apname","aplat","aplon","apmaxrwy","aptype", \
-                              "apco","apelev"]) <=    set(aptdata.keys())
-                                         
-            firdata_ok = set(["fir","firlat0","firlon0","firlat1","firlon1"])\
-                          <= set(firdata.keys())
-                          
-            codata_ok = set(["coname","cocode2","cocode3","conr"]) \
-                             <= set(codata.keys())
-                         
-            # If any of data not ok, cache is not ok
-            cache_ok = wpdata_ok and apdata_ok and firdata_ok and codata_ok     
-
-            # If not delete cache file to trigger rewriting it
-            if not cache_ok:
+            if not (cache_ok and cache_complete):
                 os.remove(cachedir + '/navdata.p')
                 print "Cache out of date: Removing old cache-file navdata.p"
 
@@ -159,8 +152,10 @@ def load_navdata():
         else:
             rwythresholds = pygame_load_rwythresholds()
             with open(cachedir + '/rwythresholds.p', 'wb') as f:
+                print "Writing cache:rwythresholds.p"
                 pickle.dump(rwythresholds,  f, pickle.HIGHEST_PROTOCOL)
     
     with open(cachedir + '/rwythresholds.p', 'rb') as f:
+        print "Reading cache: rwythresholds.p"
         rwythresholds = pickle.load(f)
     return wptdata, aptdata, firdata, codata, rwythresholds
